Combat and UI

+++ Assets/Scripts/IDamagable.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    float MaxHealth { get; set; }
-    float CurrentHealth { get; set; }
+    int MaxHealth { get; set; }
+    int CurrentHealth { get; set; }
     float Block { get; set; }
     
 
-    void TakeDamage(float amount);
+    void TakeDamage(int amount);
     void Die();
 }
